requirements.txt
# APEX F1 Intelligence Dashboard — Backend Requirements
# Install: pip install -r requirements.txt

# FastAPI
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
python-multipart==0.0.9

# Firebase Admin
firebase-admin==6.5.0

# ML
xgboost==2.0.3
scikit-learn==1.5.0
shap==0.45.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# F1 Data
fastf1==3.3.9


# Utilities
python-dotenv==1.0.1
aiohttp==3.9.5

# Development
pytest==8.2.0
httpx==0.27.0  # for test client
